+++ app/models/draft-registration.ts
@@ -2,12 +2,27 @@ import DS from 'ember-data';
 
 import { RegistrationResponse } from 'ember-osf-web/packages/registration-schema';
 
-import NodeModel from './node';
+import RegistrationProviderModel from 'ember-osf-web/models/registration-provider';
+import InstitutionModel from './institution';
+import LicenseModel from './license';
+import NodeModel, { NodeCategory, NodeLicense } from './node';
 import OsfModel from './osf-model';
 import RegistrationSchemaModel, { RegistrationMetadata } from './registration-schema';
+import SubjectModel from './subject';
 import UserModel from './user';
 
-const { attr, belongsTo } = DS;
+const { attr, belongsTo, hasMany } = DS;
+
+export enum DraftMetadataProperties {
+    Title = 'title',
+    Description = 'description',
+    Category = 'category',
+    AffiliatedInstitutions = 'affiliatedInstitutions',
+    License = 'license',
+    NodeLicenseProperty = 'nodeLicense',
+    Subjects = 'subjects',
+    Tags = 'tags',
+}
 
 export default class DraftRegistrationModel extends OsfModel {
     @attr('fixstring') registrationSupplement!: string;
@@ -15,6 +30,12 @@ export default class DraftRegistrationModel extends OsfModel {
     @attr('registration-responses') registrationResponses!: RegistrationResponse;
     @attr('date') datetimeInitiated!: Date;
     @attr('date') datetimeUpdated!: Date;
+
+    @attr('fixstring') title!: string;
+    @attr('fixstring') description!: string;
+    @attr('fixstringarray') tags!: string[];
+    @attr('node-license') nodeLicense!: NodeLicense | null;
+    @attr('node-category') category!: NodeCategory;
 
     @belongsTo('node', { inverse: 'draftRegistrations' })
     branchedFrom!: DS.PromiseObject<NodeModel> & NodeModel;
@@ -24,6 +45,18 @@ export default class DraftRegistrationModel extends OsfModel {
 
     @belongsTo('registration-schema', { inverse: null })
     registrationSchema!: DS.PromiseObject<RegistrationSchemaModel> & RegistrationSchemaModel;
+
+    @belongsTo('registration-provider', { inverse: null })
+    provider!: DS.PromiseObject<RegistrationProviderModel> & RegistrationProviderModel;
+
+    @hasMany('institution', { inverse: null, async: true })
+    affiliatedInstitutions!: DS.PromiseManyArray<InstitutionModel>;
+
+    @hasMany('subject', { inverse: null, async: true })
+    subjects!: DS.PromiseObject<SubjectModel[]> & SubjectModel[];
+
+    @belongsTo('license', { inverse: null, async: true })
+    license!: DS.PromiseObject<LicenseModel> & LicenseModel;
 }
 
 declare module 'ember-data/types/registries/model' {