+++ app/packages/registration-schema/validations.ts
@@ -1,12 +1,38 @@
 import { assert } from '@ember/debug';
+import { set } from '@ember/object';
 import { ValidationObject, ValidatorFunction } from 'ember-changeset-validations';
 import { validatePresence } from 'ember-changeset-validations/validators';
+
+import DraftRegistration, { DraftMetadataProperties } from 'ember-osf-web/models/draft-registration';
 import NodeModel from 'ember-osf-web/models/node';
 import { RegistrationResponse } from 'ember-osf-web/packages/registration-schema';
 import { SchemaBlockGroup } from 'ember-osf-web/packages/registration-schema/schema-block-group';
 import { validateFileList } from 'ember-osf-web/validators/validate-response-format';
 
-// TODO: find a way to use intl to translate error messages
+function getErrorType(groupType?: string) {
+    let validationErrorType = 'blank';
+    switch (groupType) {
+    case 'contributors-input':
+        // No validation for contributors input.
+        break;
+    case 'short-text-input':
+        break;
+    case 'long-text-input':
+        break;
+    case 'file-input':
+        validationErrorType = 'mustSelectFileMinOne';
+        break;
+    case 'single-select-input':
+        validationErrorType = 'mustSelect';
+        break;
+    case 'multi-select-input':
+        validationErrorType = 'mustSelectMinOne';
+        break;
+    default:
+        break;
+    }
+    return validationErrorType;
+}
 
 export function buildValidation(groups: SchemaBlockGroup[], node?: NodeModel) {
     const ret: ValidationObject<RegistrationResponse> = {};
@@ -17,29 +43,10 @@ export function buildValidation(groups: SchemaBlockGroup[], node?: NodeModel) {
             const responseKey = group.registrationResponseKey;
             assert(`no response key for group ${group.schemaBlockGroupKey}`, Boolean(responseKey));
             const { inputBlock } = group;
-            let type = 'blank';
-            switch (group.groupType) {
-            case 'contributors-input':
-                // No validation for contributors input.
-                return;
-            case 'short-text-input':
-                break;
-            case 'long-text-input':
-                break;
-            case 'file-input':
-                type = 'mustSelectFileMinOne';
+            if (group.groupType === 'file-input') {
                 validationForResponse.push(
                     validateFileList(responseKey as string, node),
                 );
-                break;
-            case 'single-select-input':
-                type = 'mustSelect';
-                break;
-            case 'multi-select-input':
-                type = 'mustSelectMinOne';
-                break;
-            default:
-                break;
             }
             if (inputBlock.required) {
                 validationForResponse.push(
@@ -48,7 +55,7 @@ export function buildValidation(groups: SchemaBlockGroup[], node?: NodeModel) {
                         ignoreBlank: true,
                         allowBlank: false,
                         allowNone: false,
-                        type,
+                        type: getErrorType(group.groupType),
                     }),
                 );
             }
@@ -56,4 +63,54 @@ export function buildValidation(groups: SchemaBlockGroup[], node?: NodeModel) {
         }
     });
     return ret;
+}
+
+export function validateNodeLicense() {
+    return async (_: unknown, __: unknown, ___: unknown, changes: DraftRegistration, content: DraftRegistration) => {
+        let validateLicenseTarget = await content.license;
+        let validateNodeLicenseTarget = content.nodeLicense;
+        if (changes.license) {
+            validateLicenseTarget = changes.license;
+        }
+        if (changes.nodeLicense) {
+            validateNodeLicenseTarget = changes.nodeLicense;
+        }
+        if (!validateLicenseTarget || validateLicenseTarget.get('requiredFields').length === 0) {
+            return true;
+        }
+        const missingFieldsList: string[] = [];
+        for (const item of validateLicenseTarget.get('requiredFields')) {
+            if (!validateNodeLicenseTarget || !validateNodeLicenseTarget[item]) {
+                missingFieldsList.push(item);
+            }
+        }
+        if (missingFieldsList.length === 0) {
+            return true;
+        }
+        const missingFields = missingFieldsList.join(', ');
+        return {
+            context: {
+                type: 'node_license_missing_fields',
+                translationArgs: {
+                    missingFields,
+                },
+            },
+        };
+    };
+}
+
+export function buildMetadataValidations() {
+    const validationObj: ValidationObject<DraftRegistration> = {};
+    const notBlank: ValidatorFunction[] = [validatePresence({
+        presence: true,
+        ignoreBlank: true,
+        allowBlank: false,
+        allowNone: false,
+        type: 'blank',
+    })];
+    set(validationObj, DraftMetadataProperties.Title, notBlank);
+    set(validationObj, DraftMetadataProperties.Description, notBlank);
+    set(validationObj, DraftMetadataProperties.License, notBlank);
+    set(validationObj, DraftMetadataProperties.NodeLicenseProperty, validateNodeLicense());
+    return validationObj;
 }