+++ lib/osf-components/addon/components/debouncer/component.ts
@@ -1,0 +1,28 @@
+import Component from '@ember/component';
+import { assert } from '@ember/debug';
+import { timeout } from 'ember-concurrency';
+import { task } from 'ember-concurrency-decorators';
+
+import { layout } from 'ember-osf-web/decorators/component';
+
+import template from './template';
+
+@layout(template)
+export default class Debouncer extends Component {
+    fn!: () => void;
+    timeoutInterval: number = 500;
+
+    @task({ restartable: true })
+    doFnDebounce = task(function *(this: Debouncer) {
+        yield timeout(this.timeoutInterval);
+        try {
+            yield this.fn();
+        } catch (error) {
+            throw error;
+        }
+    });
+
+    didReceiveAttrs() {
+        assert('Debouncer needs a function input', Boolean(this.fn));
+    }
+}