+++ lib/registries/addon/routes.ts
@@ -13,6 +13,8 @@ export default buildRoutes(function() {
     this.route('drafts', { path: '/registries/drafts' }, function() {
         this.route('draft', { path: '/:id' }, function() {
             this.route('page', { path: '/:page' });
+            this.route('review');
+            this.route('metadata');
         });
     });