+++ mirage/config.ts
@@ -100,6 +100,11 @@ export default function(this: Server) {
         only: ['index', 'show', 'update'],
         path: '/draft_registrations',
     });
+    osfToManyRelationship(this, 'draft-registration', 'subjects');
+    osfToManyRelationship(this, 'draft-registration', 'affiliatedInstitutions', {
+        only: ['related', 'update', 'add', 'remove'],
+        path: '/draft_registrations/:parentID/relationships/institutions',
+    });
 
     osfResource(this, 'registration', { except: ['show', 'create'] });
     this.post('/registrations', createRegistration);