+++ tests/engines/registries/acceptance/overview/overview-test.ts
@@ -407,18 +407,19 @@ module('Registries | Acceptance | overview.overview', hooks => {
         await visit(`/${reg.id}/`);
 
         assert.dom('[data-test-edit-button="license"]').isVisible();
+
         await click('[data-test-edit-button="license"]');
 
         assert.dom('[data-test-license-edit-form]').isVisible();
-        await selectSearch('[data-test-select-license]', 'MIT');
+        await selectSearch('[data-test-power-select-dropdown]', 'MIT');
         assert.dom('.ember-power-select-options').hasText('MIT License');
-        await selectSearch('[data-test-select-license]', 'No');
+        await selectSearch('[data-test-power-select-dropdown]', 'No');
         assert.dom('.ember-power-select-options').hasText('No license');
-        await selectChoose('[data-test-select-license]', 'No license');
+        await selectChoose('[data-test-power-select-dropdown]', 'No license');
 
         await click('[data-test-save-license]');
 
-        const missingFields = t('app_components.license_picker.fields.copyrightHolders').toString();
+        const missingFields = 'copyrightHolders';
         const validationErrorMsg = t('validationErrors.node_license_missing_fields', { missingFields }).toString();
         assert.dom('.help-block').hasText(validationErrorMsg, 'validation works');